refactor: DeliveryMessageListener DIP 위반 개선
