+++ com.delivery-signal.eureka.client.delivery/src/main/java/com/delivery_signal/eureka/client/delivery/application/service/DeliveryNotificationService.java
@@ -1,6 +1,6 @@
 package com.delivery_signal.eureka.client.delivery.application.service;
 
-import com.delivery_signal.eureka.client.delivery.application.port.NotificationPort;
+import com.delivery_signal.eureka.client.delivery.application.port.out.NotificationPort;
 import com.delivery_signal.eureka.client.delivery.domain.entity.Delivery;
 import com.delivery_signal.eureka.client.delivery.domain.entity.DeliveryRouteRecords;
 import com.delivery_signal.eureka.client.delivery.domain.vo.SlackMessageDetails;
