--- com.delivery-signal.eureka.client.delivery/src/main/java/com/delivery_signal/eureka/client/delivery/infrastructure/dto/DeliveryCreateDto.java
@@ -1,18 +1,0 @@
-package com.delivery_signal.eureka.client.delivery.infrastructure.dto;
-
-import java.util.UUID;
-import lombok.Builder;
-
-@Builder
-public record DeliveryCreateDto(
-    Long requestUserId,
-    UUID orderId,
-    UUID companyId,
-    String status,
-    UUID departureHubId,
-    UUID destinationHubId,
-    String address,
-    String recipient,
-    String recipientSlackId
-) {
-}
